+++ lib/src/enum/group_at_type.dart
@@ -3,4 +3,5 @@ class GroupAtType {
   static const atMe = 1;
   static const atAll = 2;
   static const atAllAtMe = 3;
+  static const groupNotification = 4;
 }