+++ lib/src/manager/im_organization_manager.dart
@@ -92,6 +92,57 @@ class OrganizationManager {
           .then((value) =>
               Utils.toObj(value, (v) => DeptMemberAndSubDept.fromJson(v)));
 
+  /// Query department info
+  /// 查询部门信息
+  Future<DeptInfo> getDeptInfo({
+    required String departmentID,
+    String? operationID,
+  }) =>
+      _channel
+          .invokeMethod(
+              'getDepartmentInfo',
+              _buildParam({
+                'departmentID': departmentID,
+                'operationID': Utils.checkOperationID(operationID),
+              }))
+          .then((value) => Utils.toObj(value, (v) => DeptInfo.fromJson(v)));
+
+  /// Search
+  /// 搜索组织人员
+  Future<OrganizationSearchResult> searchOrganization({
+    required String keyWord,
+    bool isSearchUserName = false,
+    bool isSearchEnglishName = false,
+    bool isSearchPosition = false,
+    bool isSearchUserID = false,
+    bool isSearchMobile = false,
+    bool isSearchEmail = false,
+    bool isSearchTelephone = false,
+    int offset = 0,
+    int count = 40,
+    String? operationID,
+  }) =>
+      _channel
+          .invokeMethod(
+              'searchOrganization',
+              _buildParam({
+                'searchParams': {
+                  'keyWord': keyWord,
+                  'isSearchUserName': isSearchUserName,
+                  'isSearchEnglishName': isSearchEnglishName,
+                  'isSearchPosition': isSearchPosition,
+                  'isSearchUserID': isSearchUserID,
+                  'isSearchMobile': isSearchMobile,
+                  'isSearchEmail': isSearchEmail,
+                  'isSearchTelephone': isSearchTelephone,
+                },
+                'offset': offset,
+                'count': count,
+                'operationID': Utils.checkOperationID(operationID),
+              }))
+          .then((value) =>
+              Utils.toObj(value, (v) => OrganizationSearchResult.fromJson(v)));
+
   static Map _buildParam(Map param) {
     param["ManagerName"] = "organizationManager";
     return param;